Geography page splits on litigation feature
